backend-requirements.txt
anthropic==0.40.0
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.19
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3
json-schema==4.23.0
typing-extensions==4.12.2